requirements.txt
backoff==2.2.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
geographiclib==2.0
geopy==2.4.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.2.0
pandas==2.2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
